升级 Loco 至 1.1 并简化 Agent 配置类型
更新 SeaORM、配置模板和 AppContext 初始化方式，直接集成 Utoipa 文档路由。统一主键与 Agent 配置为 i64，移除冗余整数转换。

验证：cargo test 267 项通过、3 项原有测试忽略；cargo clippy --all-targets -- -D warnings 与格式检查通过。

+++ playground/examples/mooc-manus/server/src/views/app_config.rs
@@ -60,9 +60,9 @@ impl From<LlmConfig> for LlmConfigResponse {
 /// Request payload for updating Agent config.
 #[derive(Debug, Deserialize, ToSchema)]
 pub struct AgentConfigRequest {
-    pub max_iterations: usize,
-    pub max_retries: usize,
-    pub max_search_results: usize,
+    pub max_iterations: i64,
+    pub max_retries: i64,
+    pub max_search_results: i64,
 }
 
 impl From<AgentConfigRequest> for AgentConfig {
@@ -79,9 +79,9 @@ impl From<AgentConfigRequest> for AgentConfig {
 /// Response payload for Agent config.
 #[derive(Debug, Serialize, ToSchema)]
 pub struct AgentConfigResponse {
-    pub max_iterations: usize,
-    pub max_retries: usize,
-    pub max_search_results: usize,
+    pub max_iterations: i64,
+    pub max_retries: i64,
+    pub max_search_results: i64,
 }
 
 impl From<AgentConfig> for AgentConfigResponse {
